Move iOS model list in a more appropriate location
