+++ src/sentry/api/helpers/ios_models.py
